+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@voluespark/spark.elements",
-  "version": "0.0.24",
+  "version": "0.0.26",
   "main": "build/index.js",
   "module": "build/index.esm.js",
   "files": [
