+++ src/components/UI/Range/Range.module.css
@@ -173,7 +173,7 @@
     margin: 0.5rem 0;
 }
 
-.labelFirst {
+.labelLeft {
     justify-self: start;
 }
