+++ src/components/UI/index.ts
@@ -1,1 +1,3 @@
 export { default as Radio } from './Radio'
+export { default as Range } from './Range'
+export { default as Button } from './Button'
