--- src/components/UI/index.tsx
@@ -1,2 +1,0 @@
-
-export { default as Range } from './Range'
