+++ src/index.ts
@@ -5,3 +5,6 @@ export { default as DeprecatedForecast } from './components/DeprecatedForecast'
 export { default as ForecastTable } from './components/ForecastTable'
 export { default as InteractivePriceGraph } from './components/InteractivePriceGraph'
 export { default as Settings } from './components/Settings'
+export { default as Button } from './components/UI/Button'
+export { default as Radio } from './components/UI/Radio'
+export { default as Range } from './components/UI/Range'
