+++ stories/Internal/Button.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import Button from '../../src/components/Button'
+import Button from '../../src/components/UI/Button'
 
 const meta: Meta<typeof Button> = {
     component: Button,
